requirements.txt



djangorestframework == 3.12.2
eventlet == 0.29.1
kiteconnect == 3.8.2
numpy == 1.17.3
pandas == 1.0.3
python-rest-framework == 0.3.14
requests == 2.23.0
socketio == 0.2.1
talib == 0.1.1

